server/requirements.txt
flasgger==0.9.5
Flask==2.0.3
Flask-Babel==2.0.0
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
pika==1.3.1